+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to the "box-comment" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.1.7]
+
+-   Ability to uncomment entire blocks back to normal
+-   Use block comments rather than line blocks
+-   `box-comment.indentation` setting for pre-indentation
+-   `box-comment.length` setting for total line length
+-   `box-comment.linebreak.before` boolean flag
+-   `box-comment.linebreak.after` boolean flag
+
 ## [Unreleased]
 
-- Initial release
+-   Initial release